+++ .env
@@ -1,2 +1,2 @@
 BASE_URL="/"
-BASE_API_URL="/api/"
+BASE_API_URL="http://ec2-35-87-21-24.us-west-2.compute.amazonaws.com:8081/"
