+++ src/pages/AuthConfiguration/SignUpOptions.scss
